implemented delete permission for comment and feature

+++ src/components/Comment/Comment.jsx
@@ -1,15 +1,43 @@
 import React from 'react';
 import api from '../../utils/handleApi';
 import useComments from '../../hooks/useComments';
+import useAuth from '../../hooks/useAuth';
+import Swal from 'sweetalert2'
 
 const Comment = ({ comment, featureId }) => {
-    const { refetch } = useComments(featureId)
-    // console.log(comment)
+    const { refetch, } = useComments(featureId)
+    const { user } = useAuth();
+    const currentUserEmail = user?.email;
+
+    console.log(comment?.email, currentUserEmail)
     const handleDelete = (commentId) => {
-        api.delete(`/comment/${commentId}/${featureId}`)
-            .then(() => {
-                refetch()
-            })
+
+        Swal.fire({
+            title: "Are you sure?",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, Delete!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                // Deleting the comment using an API call
+                api.delete(`/comment/${commentId}/${featureId}`)
+                    .then(() => {
+                        refetch()
+
+                    })
+                    .catch((error) => {
+                        console.error('Error deleting comment:', error);
+                    });
+
+                // Displaying a success message after deleting the feature
+                Swal.fire({
+                    title: "Deleted Successfully!",
+                    icon: "success"
+                });
+            }
+        });
+
     };
 
     return (
@@ -20,15 +48,20 @@ const Comment = ({ comment, featureId }) => {
                 className="w-10 h-10 rounded-full object-cover mr-4"
             />
             <div className="flex-1">
-                <div className="flex items-center justify-between">
-                    <p className="text-sm font-semibold">{comment.name}</p>
-                    <button
-                        onClick={() => handleDelete(comment._id)}
-                        className="text-red-500 hover:text-red-700 focus:outline-none"
-                    >
-                        Delete
-                    </button>
-                </div>
+                {
+                    currentUserEmail === comment.email ?
+                        <div className="flex items-center justify-between">
+                            <p className="text-sm font-semibold">{comment.name}</p>
+                            <button
+                                onClick={() => handleDelete(comment._id)}
+                                className="text-red-500 hover:text-red-700 focus:outline-none"
+                            >
+                                Delete
+                            </button>
+                        </div>
+                        : <></>
+                }
+
                 <p className="text-gray-600 text-sm">{comment.comment}</p>
             </div>
         </div>
